Extract provider-turn reply orchestration out of turn_coordinator
The reply/followup loop was one of the densest hot-path sections in
turn_coordinator.rs. Moving it into a dedicated provider_turn_reply
module makes the coordinator easier to navigate while keeping the
behavior and verification surface intact.

Constraint: Preserve provider-turn followup and malformed-recovery behavior while moving only the reply orchestration slice
Rejected: Leave reply-loop logic embedded and keep extracting only support helpers | the hot-path complexity would still stay concentrated in turn_coordinator.rs
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Keep pushing provider-turn slices toward focused submodules; do not drift back toward a monolithic turn_coordinator hot path
Tested: cargo fmt --all
Tested: cargo check -p loong-app
Tested: cargo test -p loong-app provider_turn_reply_tail_phase_captures_reply_and_after_turn_context
Tested: cargo test -p loong-app provider_continuation_recovers_malformed_parse_followup_without_real_tool_call
Tested: cargo test -p loong-app resolved_provider_turn_checkpoint_preserves_safe_lane_route_provenance
Not-tested: full workspace clippy/test/all-features gate; wider provider-turn regression set beyond targeted reply-loop coverage
